requirements.txt
pyyaml
sentence-transformers
transformers
torch
scikit-learn
numpy
